% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGoertzel.R
\name{goertzel}
\alias{goertzel}
\title{Generatlized Goertzel Algorithm}
\usage{
goertzel(x, k)
}
\arguments{
\item{x}{Vector containing real-valued data.}

\item{k}{Vector of real-valued elements containing the frequency components desired,
where k = i/length(x)  (k start at 0 and should be less than length(x)).}
}
\description{
Goertzel algorithm generalized to non-integer multiples of fundamental frequency
}
\details{
fft(x)[1] = goertzel(x, 0)
}
\references{
Sysel, Petr, and Pavel Rajmic. "Goertzel algorithm generalized to non-integer multiples of
  fundamental frequency." EURASIP Journal on Advances in Signal Processing 2012.1 (2012): 1-8.
}
