% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGoertzel.R
\name{goertzel}
\alias{goertzel}
\title{Generatlized Goertzel Algorithm}
\usage{
goertzel(x, k)
}
\arguments{
\item{x}{Vector containing real-valued data.}

\item{k}{Vector containing the frequency components desired (k starts at 0).}
}
\description{
Goertzel algorithm generalized to non-integer multiples of fundamental frequency
}
\references{
Sysel, Petr, and Pavel Rajmic. "Goertzel algorithm generalized to non-integer multiples of
  fundamental frequency." EURASIP Journal on Advances in Signal Processing 2012.1 (2012): 1-8.
}
